tests/migrations: separate migration utility functions into a mixin
