+++ tests/rptest/tests/data_migrations_api_test.py
@@ -13,7 +13,6 @@ import time
 from enum import Enum
 from typing import Callable, Literal, List, TypedDict, get_type_hints
 import typing
-from requests.exceptions import ConnectionError
 from contextlib import contextmanager, nullcontext
 
 from rptest.services.admin import Admin, MigrationAction
@@ -21,17 +20,18 @@ from rptest.services.admin import OutboundDataMigration, InboundDataMigration, N
 
 import confluent_kafka as ck
 
+from ducktape.mark import matrix
+from ducktape.tests.test import TestContext
+from ducktape.utils.util import wait_until
 from rptest.services.cluster import cluster
 from rptest.services.redpanda import RedpandaService, RedpandaServiceBase, SISettings
-from ducktape.utils.util import wait_until
 from rptest.services.kgo_verifier_services import KgoVerifierConsumerGroupConsumer, KgoVerifierProducer
 from rptest.services.redpanda import SISettings
 from rptest.tests.redpanda_test import RedpandaTest
-from ducktape.tests.test import TestContext
 from rptest.clients.types import TopicSpec
 from rptest.tests.e2e_finjector import Finjector
 from rptest.clients.rpk import RpkTool, RpkException
-from ducktape.mark import matrix
+from rptest.utils.data_migrations import DataMigrationTestMixin
 import requests
 import re
 
@@ -114,7 +114,7 @@ def generate_tmptpdi_params() -> List[TmtpdiParams]:
     ]
 
 
-class DataMigrationsApiTest(RedpandaTest):
+class DataMigrationsApiTest(RedpandaTest, DataMigrationTestMixin):
     log_segment_size = 10 * 1024
 
     def __init__(self, test_context: TestContext, *args, **kwargs):
@@ -125,7 +125,7 @@ class DataMigrationsApiTest(RedpandaTest):
             cloud_storage_enable_remote_read=True,
             cloud_storage_enable_remote_write=True,
         )
-        super().__init__(test_context=test_context, *args, **kwargs)
+        RedpandaTest.__init__(self, test_context=test_context, *args, **kwargs)
         self.flaky_admin = Admin(self.redpanda, retry_codes=[503, 504])
         self.admin = Admin(self.redpanda)
         self.last_producer_id = 0
@@ -190,90 +190,6 @@ class DataMigrationsApiTest(RedpandaTest):
         return self.flaky_admin_cm(
             TransferLeadersBackgroundThread(self.redpanda, topic_name))
 
-    def get_migrations_map(self, node=None):
-        migrations = self.admin.list_data_migrations(node).json()
-        return {migration["id"]: migration for migration in migrations}
-
-    def get_migration(self, id, node=None):
-        try:
-            return self.admin.get_data_migration(id, node).json()
-        except requests.exceptions.HTTPError as e:
-            if e.response.status_code == 404:
-                return None
-            else:
-                raise
-
-    def assure_not_deletable(self, id, node=None):
-        try:
-            self.admin.delete_data_migration(id, node)
-            assert False
-        except:
-            pass
-
-    def on_all_live_nodes(self, migration_id, predicate):
-        success_cnt = 0
-        exception_cnt = 0
-        for n in self.redpanda.nodes:
-            try:
-                map = self.get_migrations_map(n)
-                self.logger.debug(f"migrations on node {n.name}: {map}")
-                list_item = map[migration_id] if migration_id in map else None
-                individual = self.get_migration(migration_id, n)
-
-                if predicate(list_item) and predicate(individual):
-                    success_cnt += 1
-                else:
-                    return False
-            except ConnectionError:
-                exception_cnt += 1
-        return success_cnt > exception_cnt
-
-    def wait_for_migration_states(self,
-                                  id: int,
-                                  states: list[str],
-                                  assure_completed_after: int = 0):
-        def migration_in_one_of_states_on_node(m):
-            if m is None:
-                return False
-            completed_at = m.get("completed_timestamp")
-            if m["state"] in ("finished", "cancelled"):
-                self.validate_timing(assure_completed_after, completed_at)
-            else:
-                assert "completed_timestamp" not in m
-            return m["state"] in states
-
-        def migration_in_one_of_states():
-            return self.on_all_live_nodes(id,
-                                          migration_in_one_of_states_on_node)
-
-        self.logger.info(f'waiting for {" or ".join(states)}')
-        wait_until(
-            migration_in_one_of_states,
-            timeout_sec=90,
-            backoff_sec=1,
-            err_msg=
-            f"Failed waiting for migration {id} to reach one of {states} states"
-        )
-        if all(state not in ('planned', 'finished', 'cancelled')
-               for state in states):
-            self.assure_not_deletable(id)
-
-    def wait_migration_appear(self, migration_id, assure_created_after):
-        def migration_present_on_node(m):
-            if m is None:
-                return False
-            self.validate_timing(assure_created_after, m['created_timestamp'])
-            return True
-
-        def migration_is_present(id: int):
-            return self.on_all_live_nodes(id, migration_present_on_node)
-
-        wait_until(
-            lambda: migration_is_present(migration_id),
-            timeout_sec=30,
-            backoff_sec=2,
-            err_msg=f"Expected migration with id {migration_id} is present")
-
     def wait_migration_disappear(self, migration_id):
         def migration_is_absent(id: int):
             return self.on_all_live_nodes(id, lambda m: m is None)
@@ -283,62 +199,6 @@ class DataMigrationsApiTest(RedpandaTest):
             timeout_sec=90,
             backoff_sec=2,
             err_msg=f"Expected migration with id {migration_id} is absent")
-
-    def wait_partitions_appear(self, topics: list[TopicSpec]):
-        # we may be unlucky to query a slow node
-        def topic_has_all_partitions(t: TopicSpec):
-            part_cnt = len(self.client().describe_topic(t.name).partitions)
-            self.logger.debug(
-                f"topic {t.name} has {part_cnt} partitions out of {t.partition_count} expected"
-            )
-            return t.partition_count == part_cnt
-
-        def err_msg():
-            msg = "Failed waiting for partitions to appear:\n"
-            for t in topics:
-                msg += f"   {t.name} expected {t.partition_count} partitions, "
-                msg += f"got {len(self.client().describe_topic(t.name).partitions)} partitions\n"
-            return msg
-
-        wait_until(lambda: all(topic_has_all_partitions(t) for t in topics),
-                   timeout_sec=90,
-                   backoff_sec=1,
-                   err_msg=err_msg)
-
-    def wait_partitions_disappear(self, topics: list[TopicSpec]):
-        # we may be unlucky to query a slow node
-        wait_until(
-            lambda: all(self.client().describe_topic(t.name).partitions == []
-                        for t in topics),
-            timeout_sec=90,
-            backoff_sec=1,
-            err_msg=f"Failed waiting for partitions to disappear")
-
-    def create_and_wait(self, migration: InboundDataMigration
-                        | OutboundDataMigration):
-        def migration_id_if_exists():
-            for n in self.redpanda.nodes:
-                for m in self.admin.list_data_migrations(n).json():
-                    if m == migration:
-                        return m[id]
-            return None
-
-        time_before_creation = now()
-        try:
-            reply = self.admin.create_data_migration(migration).json()
-            self.logger.info(f"create migration reply: {reply}")
-            migration_id = reply["id"]
-        except requests.exceptions.HTTPError as e:
-            maybe_id = migration_id_if_exists()
-            if maybe_id is None:
-                raise
-            migration_id = maybe_id
-            self.logger.info(f"create migration failed "
-                             f"but migration {migration_id} present: {e}")
-
-        self.wait_migration_appear(migration_id, time_before_creation)
-
-        return migration_id
 
     def assure_not_migratable(self, topic: TopicSpec, expected_response=None):
         out_migration = OutboundDataMigration(
@@ -563,7 +423,7 @@ class DataMigrationsApiTest(RedpandaTest):
             # still preparing, i.e. stuck
             self.wait_for_migration_states(in_migration_id, ['preparing'])
             # and the topic is not there
-            self.wait_partitions_disappear([topic])
+            self.wait_partitions_disappear([topic.name])
 
             time_before_final_action = now()
             self.execute_data_migration_action_flaky(in_migration_id,
@@ -574,7 +434,7 @@ class DataMigrationsApiTest(RedpandaTest):
             self.wait_for_migration_states(in_migration_id, ['cancelled'],
                                            time_before_final_action)
             # still not there
-            self.wait_partitions_disappear([topic])
+            self.wait_partitions_disappear([topic.name])
 
             self.admin.delete_data_migration(in_migration_id)
             self.wait_migration_disappear(in_migration_id)
@@ -653,7 +513,7 @@ class DataMigrationsApiTest(RedpandaTest):
             self.wait_for_migration_states(out_migration_id, ['finished'],
                                            time_before_final_action)
 
-            self.wait_partitions_disappear(topics)
+            self.wait_partitions_disappear([t.name for t in topics])
 
             # in
             inbound_topics = [
@@ -735,7 +595,7 @@ class DataMigrationsApiTest(RedpandaTest):
             producer.flush()
             revisions[i] = self.get_topic_initial_revision(topic.name)
             out_migr_id = self.admin.unmount_topics([ns_topic]).json()["id"]
-            self.wait_partitions_disappear([topic])
+            self.wait_partitions_disappear([topic.name])
             self.wait_migration_disappear(out_migr_id)
 
         # mount and consume from them in random order
@@ -777,7 +637,7 @@ class DataMigrationsApiTest(RedpandaTest):
         out_migration_id = reply["id"]
         with self.finj_thread():
             self.logger.info('waiting for partitions be deleted')
-            self.wait_partitions_disappear(topics)
+            self.wait_partitions_disappear([t.name for t in topics])
             self.logger.info('waiting for migration to be deleted')
             self.wait_migration_disappear(out_migration_id)
 
@@ -1227,7 +1087,7 @@ class DataMigrationsApiTest(RedpandaTest):
         self.logger.info(f"create migration reply: {reply}")
 
         self.logger.info('waiting for partitions be deleted')
-        self.wait_partitions_disappear(topics)
+        self.wait_partitions_disappear([t.name for t in topics])
 
         list_mountable_res = admin.list_mountable_topics().json()
         assert len(list_mountable_res["topics"]) == len(
